+++ Src/Sensor/kalmanfilter.h
@@ -5,16 +5,16 @@
 #include "stm32f1xx_hal_tim.h"
 
 typedef struct {
-    float q; /*< process noise covariance */
-    float r; /*< measurement noise covariance */
-    float x; /*< value */
-    float p; /*< estimation error covariance */
-    float k; /*< kalman gain */
+    double q; /*< process noise covariance */
+    double r; /*< measurement noise covariance */
+    double x; /*< value */
+    double p; /*< estimation error covariance */
+    double k; /*< kalman gain */
     uint8_t is_initialized; /*< is init or not */
 } kalman_state;
 
-void kalman_init(kalman_state* state, float q, float r, float p, float intial_value);
-void kalman_update(kalman_state* state, float measurement);
+void kalman_init(kalman_state* state, double q, double r, double p, double intial_value);
+void kalman_update(kalman_state* state, double measurement);
 
 
 #endif /* __KALMANFILTER_H__ */
